+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/combat/SuperKnockback.kt
@@ -26,8 +26,8 @@ import kotlin.math.abs
 
 object SuperKnockback : Module("SuperKnockback", ModuleCategory.COMBAT, hideModule = false) {
 
-    private val delay by IntegerValue("Delay", 0, 0, 500)
-    private val hurtTime by IntegerValue("HurtTime", 10, 0, 10)
+    private val delay by IntegerValue("Delay", 0, 0..500)
+    private val hurtTime by IntegerValue("HurtTime", 10, 0..10)
 
     private val mode by ListValue("Mode", arrayOf("SprintTap", "SprintTap2", "WTap", "Old", "Silent", "Packet", "SneakPacket"), "Old")
     private val maxTicksUntilBlock: IntegerValue = object : IntegerValue("MaxTicksUntilBlock", 2, 0..5) {
