+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/exploit/OldPhase.kt
@@ -1,0 +1,80 @@
+/*
+ * LiquidBounce Hacked Client
+ * A free open source mixin-based injection hacked client for Minecraft using Minecraft Forge.
+ * https://github.com/CCBlueX/LiquidBounce/
+ */
+package net.ccbluex.liquidbounce.features.module.modules.exploit
+
+import net.ccbluex.liquidbounce.event.EventTarget
+import net.ccbluex.liquidbounce.event.*
+import net.ccbluex.liquidbounce.features.module.Module
+import net.ccbluex.liquidbounce.features.module.ModuleCategory
+import net.ccbluex.liquidbounce.utils.MovementUtils.direction
+import net.ccbluex.liquidbounce.utils.MovementUtils.isMoving
+import net.ccbluex.liquidbounce.utils.PacketUtils.sendPackets
+import net.ccbluex.liquidbounce.utils.block.BlockUtils.collideBlockIntersects
+import net.ccbluex.liquidbounce.utils.extensions.getBlock
+import net.ccbluex.liquidbounce.utils.extensions.step
+import net.ccbluex.liquidbounce.utils.timing.TickTimer
+import net.ccbluex.liquidbounce.value.ListValue
+import net.minecraft.block.Block
+import net.minecraft.init.Blocks.air
+import net.minecraft.network.play.client.C03PacketPlayer
+import net.minecraft.network.play.client.C03PacketPlayer.C04PacketPlayerPosition
+import net.minecraft.util.AxisAlignedBB
+import net.minecraft.util.BlockPos
+import kotlin.math.cos
+import kotlin.math.sin
+
+object Phase : Module("Phase", ModuleCategory.EXPLOIT) {
+    private val mode by ListValue("Mode", arrayOf("Vanilla", "Skip", "Spartan", "Clip", "AAC3.5.0", "Mineplex", "FullBlock"), "Vanilla")
+
+    private val tickTimer = TickTimer()
+    private var shouldContinue = false
+    private var clipState = 0
+    private var yaw = 0.0
+    private var value = 0.0
+
+    @EventTarget
+    fun onUpdate(event: UpdateEvent) {
+        if (mode == "FullBlock") {
+            return
+        }
+
+        when (mode) {
+            "Spartan" -> { 
+                if (mc.thePlayer.onGround && tickTimer.hasTimePassed(2) && mc.thePlayer.isCollidedHorizontally && (!isInsideBlock || mc.thePlayer.isSneaking)) {
+                    sendPackets(
+                        C04PacketPlayerPosition(mc.thePlayer.posX, mc.thePlayer.posY, mc.thePlayer.posZ, true),
+                        C04PacketPlayerPosition(0.5, 0.0, 0.5, true),
+                        C04PacketPlayerPosition(mc.thePlayer.posX, mc.thePlayer.posY, mc.thePlayer.posZ, true),
+                        C04PacketPlayerPosition(mc.thePlayer.posX, mc.thePlayer.posY - 0.2, mc.thePlayer.posZ, true),
+                        C04PacketPlayerPosition(0.5, 0.0, 0.5, true),
+                        C04PacketPlayerPosition(mc.thePlayer.posX + 0.5, mc.thePlayer.posY, mc.thePlayer.posZ + 0.5, true)
+                    )
+
+                    val yaw = Math.toRadians(mc.thePlayer.rotationYaw.toDouble())
+                    val x = -sin(yaw) * 0.04
+                    val z = cos(yaw) * 0.04
+
+                    mc.thePlayer.setPosition(mc.thePlayer.posX + x, mc.thePlayer.posY, mc.thePlayer.posZ + z)
+                    tickTimer.reset()
+                }
+            }
+        }
+        tickTimer.update()
+    }
+
+    override fun onEnable() {
+        shouldContinue = false
+        clipState = 0
+        value = 0.0
+    }
+
+    @EventTarget
+    fun onPushOut(event: PushOutEvent) {
+        event.cancelEvent()
+    }
+
+    override val tag get() = mode
+}
