+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/exploit/phasemodes/mineplex/Mineplex.kt
@@ -1,0 +1,40 @@
+/*
+ * LiquidBounce Hacked Client
+ * A free open source mixin-based injection hacked client for Minecraft using Minecraft Forge.
+ * https://github.com/CCBlueX/LiquidBounce/
+ */
+package net.ccbluex.liquidbounce.features.module.modules.exploit.phasemodes.mineplex
+
+import net.ccbluex.liquidbounce.event.MoveEvent
+import net.ccbluex.liquidbounce.features.module.modules.exploit.Phase.tickTimer
+import net.ccbluex.liquidbounce.features.module.modules.exploit.phasemodes.PhaseMode
+import net.ccbluex.liquidbounce.utils.timing.TickTimer
+
+object Mineplex : PhaseMode("Mineplex") {
+    private var mineplexClip = false
+    private val mineplexTickTimer = TickTimer()
+
+    override fun onMove(event: MoveEvent) {
+        if (mc.thePlayer.isCollidedHorizontally) {
+            mineplexClip = true
+
+            if (!mineplexClip)
+                return
+
+            mineplexTickTimer.update()
+
+            event.x = 0.0
+            event.z = 0.0
+
+            if (mineplexTickTimer.hasTimePassed(3)) {
+                mineplexTickTimer.reset()
+                mineplexClip = false
+            } else if (mineplexTickTimer.hasTimePassed(1)) {
+                val offset = if (mineplexTickTimer.hasTimePassed(2)) 1.6 else 0.06
+                val direction = direction
+
+                mc.thePlayer.setPosition(mc.thePlayer.posX + (-sin(direction) * offset), mc.thePlayer.posY, mc.thePlayer.posZ + (cos(direction) * offset))
+            }
+        }
+    }
+}
