+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/exploit/phasemodes/other/Clip.kt
@@ -1,0 +1,32 @@
+/*
+ * LiquidBounce Hacked Client
+ * A free open source mixin-based injection hacked client for Minecraft using Minecraft Forge.
+ * https://github.com/CCBlueX/LiquidBounce/
+ */
+package net.ccbluex.liquidbounce.features.module.modules.exploit.phasemodes.other
+
+import net.ccbluex.liquidbounce.event.UpdateEvent
+import net.ccbluex.liquidbounce.features.module.modules.exploit.Phase.tickTimer
+import net.ccbluex.liquidbounce.features.module.modules.exploit.phasemodes.PhaseMode
+
+object Clip : PhaseMode("Clip") {
+	
+	override fun onUpdate(event: UpdateEvent) {
+        if (tickTimer.hasTimePassed(2) && mc.thePlayer.isCollidedHorizontally && (!isInsideBlock || mc.thePlayer.isSneaking)) {
+            val yaw = Math.toRadians(mc.thePlayer.rotationYaw.toDouble())
+            val oldX = mc.thePlayer.posX
+            val oldZ = mc.thePlayer.posZ
+
+            for (i in 1..10) {
+                val x = -sin(yaw) * i
+                val z = cos(yaw) * i
+
+                if (BlockPos(oldX + x, mc.thePlayer.posY, oldZ + z).getBlock() === air && BlockPos(oldX + x, mc.thePlayer.posY + 1, oldZ + z).getBlock() === air) {
+                    mc.thePlayer.setPosition(oldX + x, mc.thePlayer.posY, oldZ + z)
+                }
+            }
+            tickTimer.reset()
+        }
+	}
+}
+
