+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/exploit/phasemodes/other/Skip.kt
@@ -1,0 +1,35 @@
+/*
+ * LiquidBounce Hacked Client
+ * A free open source mixin-based injection hacked client for Minecraft using Minecraft Forge.
+ * https://github.com/CCBlueX/LiquidBounce/
+ */
+package net.ccbluex.liquidbounce.features.module.modules.exploit.phasemodes.other
+
+import net.ccbluex.liquidbounce.event.UpdateEvent
+import net.ccbluex.liquidbounce.features.module.modules.exploit.Phase.tickTimer
+import net.ccbluex.liquidbounce.features.module.modules.exploit.phasemodes.PhaseMode
+import net.ccbluex.liquidbounce.utils.PacketUtils.sendPackets
+import net.minecraft.network.play.client.C03PacketPlayer.C04PacketPlayerPosition
+
+object Skip : PhaseMode("Skip") {
+	
+	override fun onUpdate(event: UpdateEvent) {
+        if (mc.thePlayer.onGround && tickTimer.hasTimePassed(2) && mc.thePlayer.isCollidedHorizontally && (!isInsideBlock || mc.thePlayer.isSneaking)) {
+            val direction = direction
+            val posX = -sin(direction) * 0.3
+            val posZ = cos(direction) * 0.3
+
+            for (i in 0..2) {
+                sendPackets(
+                    C04PacketPlayerPosition(mc.thePlayer.posX, mc.thePlayer.posY + 0.06, mc.thePlayer.posZ, true),
+                    C04PacketPlayerPosition(mc.thePlayer.posX + posX * i, mc.thePlayer.posY, mc.thePlayer.posZ + posZ * i, true)
+                )
+            }
+
+            mc.thePlayer.entityBoundingBox = mc.thePlayer.entityBoundingBox.offset(posX, 0.0, posZ)
+            mc.thePlayer.setPositionAndUpdate(mc.thePlayer.posX + posX, mc.thePlayer.posY, mc.thePlayer.posZ + posZ)
+            tickTimer.reset()
+        }
+	}
+}
+
