+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/exploit/phasemodes/spartan/Spartan.kt
@@ -1,0 +1,34 @@
+/*
+ * LiquidBounce Hacked Client
+ * A free open source mixin-based injection hacked client for Minecraft using Minecraft Forge.
+ * https://github.com/CCBlueX/LiquidBounce/
+ */
+package net.ccbluex.liquidbounce.features.module.modules.exploit.phasemodes.spartan
+
+import net.ccbluex.liquidbounce.event.UpdateEvent
+import net.ccbluex.liquidbounce.features.module.modules.exploit.Phase.tickTimer
+import net.ccbluex.liquidbounce.features.module.modules.exploit.phasemodes.PhaseMode
+import net.ccbluex.liquidbounce.utils.PacketUtils.sendPackets
+import net.minecraft.network.play.client.C03PacketPlayer.C04PacketPlayerPosition
+
+object Spartam : PhaseMode("Spartan") {
+    override fun onUpdate(event: UpdateEvent) {
+        if (mc.thePlayer.onGround && tickTimer.hasTimePassed(2) && mc.thePlayer.isCollidedHorizontally && (!isInsideBlock || mc.thePlayer.isSneaking)) {
+            sendPackets(
+                C04PacketPlayerPosition(mc.thePlayer.posX, mc.thePlayer.posY, mc.thePlayer.posZ, true),
+                C04PacketPlayerPosition(0.5, 0.0, 0.5, true),
+                C04PacketPlayerPosition(mc.thePlayer.posX, mc.thePlayer.posY, mc.thePlayer.posZ, true),
+                C04PacketPlayerPosition(mc.thePlayer.posX, mc.thePlayer.posY - 0.2, mc.thePlayer.posZ, true),
+                C04PacketPlayerPosition(0.5, 0.0, 0.5, true),
+                C04PacketPlayerPosition(mc.thePlayer.posX + 0.5, mc.thePlayer.posY, mc.thePlayer.posZ + 0.5, true)
+            )
+
+            val yaw = Math.toRadians(mc.thePlayer.rotationYaw.toDouble())
+            val x = -sin(yaw) * 0.04
+            val z = cos(yaw) * 0.04
+
+            mc.thePlayer.setPosition(mc.thePlayer.posX + x, mc.thePlayer.posY, mc.thePlayer.posZ + z)
+            tickTimer.reset()
+        }
+    }
+}
